Add descriptions for new methods added and also move couch view accessor method to dbaccessors [skip ci].
